moved all js files to src directory except index.js
