+++ homework-02/index.js
@@ -1,7 +1,7 @@
-const add = require('./operations/add.js');
-const div = require('./operations/div');
-const mul = require('./operations/mul');
-const sub = require('./operations/sub');
+const add = require('./src/operations/add.js');
+const div = require('./src/operations/div');
+const mul = require('./src/operations/mul');
+const sub = require('./src/operations/sub');
 
 module.exports = {
     add,
