+++ homework-02/package.json
@@ -4,7 +4,7 @@
   "description": "Simple arithmetic operations",
   "main": "index.js",
   "bin": {
-    "calculate": "./cli.js"
+    "calculate": "src/cli.js"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
