+++ homework-02/src/errorHandler.js
