+++ homework-02/src/operations/mul.js
